pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Template Python App"
authors = ["Your Name Here <yourname@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/..."
repository = "https://github.com/..."

[tool.poetry.dependencies]
python = "^3.10.0"
websockets = "^12.0"
twitchio = "^2.10.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.12.0"
flake8 = "^6.0.0"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^22.1.11"
flake8-docstrings = "^1.6.0"
flake8-isort = "^6.0.0"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"
pep8-naming = "^0.13.3"
pre-commit = "^2.13.0"
taskipy = "^1.8.1"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100

[tool.taskipy.tasks]
app = { cmd = "python -m app", help = "Runs the main package"}
lint = { cmd = "pre-commit run --all-files", help = "Lints project files" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black --target-version py310 .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
